P0026G-4 task 8a player screens

Approved-by: Musaddiq Ansar

+++ app/src/main/java/com/techswivel/qthemusic/models/Subscription.kt
@@ -9,7 +9,7 @@ data class Subscription(
     @SerializedName("planTitle")
     var planTitle: String?,
     @SerializedName("planPrice")
-    var planPrice: Int?,
+    var planPrice: Float?,
     @SerializedName("planDuration")
     var planDuration: String?
 ) : Serializable
